Updates signing tests and documentation * Breaks signing tests into individual files * Splits ETH tx and ABI test scripts * Updates signing and testing documentation
