+++ src/types/firmware.d.ts
@@ -23,7 +23,6 @@ interface GenericSigningData {
   encodingTypes: {
     NONE: EXTERNAL.SIGNING.ENCODINGS.NONE;
     SOLANA: EXTERNAL.SIGNING.ENCODINGS.SOLANA;
-    TERRA?: EXTERNAL.SIGNING.ENCODINGS.TERRA;
     EVM?: EXTERNAL.SIGNING.ENCODINGS.EVM;
   };
 }
